% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delta_SEIR_cpp}
\alias{delta_SEIR_cpp}
\title{Calculate the SEIR change based on the event rate.}
\usage{
delta_SEIR_cpp(event)
}
\arguments{
\item{event}{The numeric vector from event_rate function.}
}
\value{
A vector of SEIR change.
}
\description{
Calculate the SEIR change based on the event rate.
}
