% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{event_rate_coloc_cpp}
\alias{event_rate_coloc_cpp}
\title{Calculate the event rate of contact model based on SEIR vector.}
\usage{
event_rate_coloc_cpp(SEIR_vector, numpatch, Di, De, R0_matrix, N)
}
\arguments{
\item{SEIR_vector}{The numeric vector of current SEIR status.}

\item{numpatch}{length of location list.}

\item{Di}{The duration of infectious period (days).}

\item{De}{The duration of exposed period (days).}

\item{R0_matrix}{The matrix of R0 between two locations.}

\item{N}{The vector of population at each location.}
}
\value{
A vector of event rate.
}
\description{
Calculate the event rate of contact model based on SEIR vector.
}
