% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{event_rate_move_cpp}
\alias{event_rate_move_cpp}
\title{Calculate the event rate of residence model based on SEIR vector.}
\usage{
event_rate_move_cpp(SEIR_vector, numpatch, Di, De, R0, N, P)
}
\arguments{
\item{SEIR_vector}{The numeric vector of current SEIR status.}

\item{numpatch}{length of location list.}

\item{Di}{The duration of infectious period (days).}

\item{De}{The duration of exposed period (days).}

\item{R0}{The vector of R0 in each location.}

\item{N}{The vector of population at each location.}

\item{P}{The matrix of possibility that a resident in location i move to locaiton j in a unit of time.}
}
\value{
A vector of event rate.
}
\description{
Calculate the event rate of residence model based on SEIR vector.
}
