% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSEIR.R
\name{simulateSEIR}
\alias{simulateSEIR}
\title{Simple stochastic SEIR simulation model with travel reduction}
\usage{
simulateSEIR(
  dataMatrix,
  locationList,
  populationList,
  initialLoc,
  model = c("residence", "contact"),
  reduceMethod = c("all", "intercity", "intracity"),
  Di,
  De,
  nsim = 1000,
  R0_ref,
  R0_max = 3.5,
  caseThreshold = 1000,
  susMonth = 3,
  decline = 0.5,
  initial.I = 5,
  threads = 1,
  seed = 1
)
}
\arguments{
\item{dataMatrix}{Residence or contact data matrix for simulation.}

\item{locationList}{The character vector of locations which involved in the spreading progress.}

\item{populationList}{The numeric vector of populations of each location in the locationList vector.}

\item{initialLoc}{The location of initial infectious starts (character).}

\item{model}{The simulation model. Must be "residence" or "contact".}

\item{reduceMethod}{The method which will be used in travel reduction. Must be "intercity", "intracity", or "all".}

\item{Di}{The duration of infectious period (days).}

\item{De}{The duration of exposed period (days).}

\item{nsim}{Number of simulations. Default is 1000.}

\item{R0_ref}{The base reference R0.}

\item{R0_max}{The maximum value of R0. Default is 3.5.}

\item{caseThreshold}{The threshold that stops the simulation process when the number of infectious exceed it. Default is 1000.}

\item{susMonth}{The time period of travel reduction (month). Default is 3.}

\item{decline}{A 0 to 1 numeric variable which represents the ratio of reduction. 0 means completely reduction and 1 means no reduction. Default is 0.5.}

\item{initial.I}{The number of initial infectious number. Default is 5.}

\item{threads}{A parallel computation option, setting the number of threads this function will make. Default is 1 (no parallel computing).}

\item{seed}{Random seed. Default is 1.}
}
\value{
A list that contains three components: \itemize{  
\item{probability}{: the probability that the number of infectious exceeds the threshold.} 
\item{accumulate_case}{: the accumulated number of infectious in each location at the end of each simulation.} 
\item{spend_day}{: the days costs at the end of each simulation.}
}
}
\description{
A SEIR simulation tool using residence or contact model.
}
\details{
This tool mainly use two models: residence model and contact model. The data input of the former model is a matrix that 
contains the directional movement of residents in each residental regions. The data matrix of contact model contains the 
value of contact possibilities between any two residents from their residental regions.
}
\examples{
location_list = c("loc1", "loc2", "loc3", "loc4", "loc5")
loc_N = c(5382, 6609, 7193, 4572, 7760)
movMat = t(matrix(c(4445, 32, 411, 132, 362, 
                    223, 5441, 541, 117, 287, 
                    443, 729, 4969, 650, 402, 
                    201, 387, 346, 3456, 182, 
                    398, 421, 1282, 571, 5088), ncol = 5))
iniLoc = "loc3"
Rmatrix = fitMoveMatrix(movMat, location_list, loc_N, alpha_F = 1000)[["Sij"]]
result = simulateSEIR(Rmatrix, locationList = location_list, populationList = loc_N, initialLoc = iniLoc, 
                      model = "residence", Di = 3, De = 3.5, R0_ref = 2.4, threads = 4)
}
\keyword{SEIR}
